python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
mysql-connector-python==8.2.0
redis==5.0.1
openai==1.3.6
python-dotenv==1.0.0
pydantic==2.5.0
maturin==1.4.0
google-auth-oauthlib==1.1.0
google-auth==2.25.2
google-api-python-client==2.110.0 